requirements.txt
neo4j>=5.14.0
networkx>=3.2.1
apache-airflow>=2.10.1
psycopg2>=2.9.10
